Group `Text` heading examples
